Fix combination of oss and release version
There is no `oss` version of Elastic Stack after version 7.x. So we need
to make sure we don't run tests for variants that don't even exist.

That can mean to pin tests to a certain release. Meaning we run the same
test twice which is a bit of a waste of ressources but will only be
neccessary as long as 7.x is supported. Or changing the variant to
`elastic` where appropriate.