Show more meaningful errormessages for API (#118)
The current variant might not be very helpful when it comes to
errormessages. But it will work better with slow setups.

We *could* use the code I removed (see commits in between, there were
also more checks) and just don't fail with the API query. I'm open for
suggestions. Personally I like the current variant - it's just sleeker.

fixes #100
fixes #95