Make an linter exception for no-handler
We're moving a lot of files during backups of certificates. Ansible Lint
thinks this should be handlers but we need the changes to be done before
the next tasks run. And triggering all handlers each time doesn't sound
like a good idea.

I guess, we'll replace some of the certificate handling with a module
later so we don't need to be over exact when making this exception. It
would be better to move all tasks to handlers where it makes sense and
make local exceptions for the ones that don't work differently, though.
That's why I leave them as "warning" and not just ignore them.