Rework automated tests (#122)
We have lots and lots of tests that stem from the elastic stack roles we
had before this collection. While developing the collection further we
found that some tests are now redundant, others are useless now and even
more others are missing.

So we take the chance to rework our automated checks as much as
neccessary to get consistent checks that won't take forever to run for a
simple small change.

fixes #120
fixes #122 through the remove of molecule full stack scenarios in beats,
logstash and kibana related GitHub workflows.