Change used module from 'command' to 'shell' to allow shell redirect (#200)
Fix #198

This PR adjusts a task that used the `command` module with a command
that wants to redirect output into a file.
This, however, is not possible since no actual shell is used, thus
making that redirect impossible. The Ingest Pipelines were never
created.
Changed the `command` module to the `shell` module.

Co-authored-by: Afeef Ghannam <39904920+afeefghannam89@users.noreply.github.com>
Co-authored-by: Thomas Widhalm <thomas.widhalm@netways.de>