Fix/lint workflow 192 (#199)
Fix #192.

This PR introduces a reusable workflow. This general workflow handles
the job necessary to run both `ansible-lint` and `yamllint` on
**either** all roles or one specific role.

All roles are linted when:
- linting rules are changed
- the linting workflow is changed

Specific roles are linted when:
- a role specific workflow is triggered (linting as first job)
- the `full_stack` workflow is triggered (linting as first job)

The workflow for the specific role now uses a call to the general
linting workflow, while providing it with its own role name.

This approach does not completely remove linting from single workflows,
but it makes those linting jobs much more readable and reduces redundant
code, while also allowing for a bit more flexibility.

---------

Co-authored-by: Thomas Widhalm <thomas.widhalm@netways.de>