Add code for automatic updates
Adding playbook contributed by @xtruthx

I have to admit, I kept this playbook for waaaaay too long. @xtruthx
sent it to me a long while back and I always planned to integrate it
into the codebase of this collection.

Now I had to face that I took too long and so I'm putting it up publicly
so we can work on integration as a common effort.

I left the playbook "as is". It stems from a different project so
variable names etc. don't match. We need to fix them first. Also I'm not
sure if we should really add it as a playbook or maybe better make it a
task file in a role.

fixes #216