Use common variable for TLS
fixes #248

This is something between a workaround and a partial solution. While it
makes more sense to use a role related variable for both Logstash and
Elasticsearch output (other than before) it's still not the goal we
want.

So overhauling the whole process of determining whether we need TLS or
not is due. I'll start a discussion and we need to adjust all roles to
the solution we find. Now there are similarities but it's not exactly
the same in every role.