Remove actions for pushs  (#303)
At the moment workflows are triggered for PRs and pushs.

The workflow to contribute says that a PR must be created beforehand for
each contribution into main.
For this reason (and to avoid duplicate workflows), github workflows
should only be triggered for PRs.

I have also checked this branch before creating the PR. No workflow were
scheduled.

Fixes #302

---------

Co-authored-by: Thomas Widhalm <thomas.widhalm@netways.de>