Enforce repos only with full stack (#314)
RPM installation automatically enabled the Elastic Stack repositories
this collection will add to a host. But if you have another tool to
manage your repositories, they might have other ids and names.

So I added separate installation tasks. If `elasticstack_full_stack` is
set to `true`, everything stays as it was. The task will forcefully
enable the repository and then install the package.

But if `elasticstack_full_stack` is set to `false` it will just ignore
the repository and expect the package to be available.

While working on this I found a different approach to choosing the
correct version in the `beats` role. I couldn't afford to fix it all in
a single PR so I opened #313 for this.

fixes #312