Use elasticsearch_network_host for connection tests if defined (#315)
When `elasticsearch_network_host` is set to something not including
localhost, all connection tests fails.

This uses `elasticsearch_network_host` for connections if it is defined,
and falls back to localhost if not.

(There might be more places this change should be made, but the role
works for me with this fix)

---------

Co-authored-by: Ivar Eriksen <ivar.eriksen@norkart.no>