Add code for automatic updates (#250)
Adding playbook contributed by @xtruthx

I have to admit, I kept this playbook for waaaaay too long. @xtruthx
sent it to me a long while back and I always planned to integrate it
into the codebase of this collection.

Now I had to face that I took too long and so I put it up publicly. Now
I worked it over so that it will fit into the `elasticsearch` role just
well.

This PR also introduces a new role where all "global" variables can be
found. This will help determining the correct version of a package to
install. Some tasks that were repeated in every role get moved into this
role called `elasticstack` to streamline execution.

This PR now reads the current version of Elasticsearch installed on the
CA host and uses it as new installation target for all further
installations. This not only helps with upgrading, it will also make
sure that you won't have a version mismatch if you install components
after the initial install.

Special thanks to @hryamzik for the workaround to run a task file like
with `serial: 1`. Found in
https://github.com/ansible/ansible/issues/12170#issuecomment-372263149

Very special thanks to CID GmbH for sponsoring parts of the the
development of this PR.


fixes #216
fixes #286 
fixes #236
fixes #205

---------

Co-authored-by: Daniel Neuberger <daniel.neuberger@netways.de>