Tune down Ansible Verbosity (#337)
While sometimes it's really helpful to have a high verbosity, most of
the time it's just overwhelming or plain annoying.

I haven't found a way to use a variable to set verbosity on a more
global space. So I left in the code but as a comment. Please feel free
to provide an alternative solution so we can tune it more elegantly.