Fix Python module installation (#336)
There are several ways of installing Python modules on a system. We
can't make safe assumptions about the system where the collection is
used.

Possible problems:
* Modules available via package manager can be too old
* Python could be entirely managed via package manager and pip wouldn't
break that
* Pip could be installed in different ways

This change will do the following:

* Install `pip` via package manager if a new variable is set
* Try to install modules via pip
* If that fails because Python is managed via package manager, try
installing them via package manager
* One of the above methods should be successful but you still might end
up with modules which are too old to work
* If that's the case you can force the installation via pip and break
the managed Python installation

fixes #334