Fix current problems with molecule checks (#360)
There are two problems that both break tests. I try to solve them in a
single PR because one would block merging the other. Both are "new
problems" that arose through updates of tools being used. The checks
worked in the past.

* Having `become` only in tasks where needed might be good coding style
but it breaks tests with Rocky Linux and might have other unpredicted
side effects. So I'm removing it.
* Starting `rsyslog` on Rocky Linux fails with "Timeout". There' no
further information so I decided to replace it with reading journald
instead of rsyslog. This might have been a move to make in the future
anyway.