Pin version of Python elasticsearch client (#391)
There seems to be no version of the Python `elasticsearch` client that
is compatible to all versions. It's either version 9 or below 9. So for
now we pin the version of the client to one that is compatible to
versions below 9. We need to take care of that when supporting Elastic
Stack 9.