Use common variable for TLS (#274)
fixes #248

The `metricbeat.yml` and `auditbeat.yml` templates have code for both
connecting to Logstash or Elasticseach. They used different variables to
determine whether to turn on TLS or not. That resulted in a missing
default and therefore Beats setup breaking when connecting to
Elasticseach. I changed the variable that's queried to determine whether
to turn on TLS or not.

This is something between a workaround and a partial solution. While it
makes more sense to use a role related variable for both Logstash and
Elasticsearch output (other than before) it's still not the goal we
want.

So overhauling the whole process of determining whether we need TLS or
not is due. I'll start a discussion and we need to adjust all roles to
the solution we find. Now there are similarities but it's not exactly
the same in every role.