Implement global handlers
There are tasks that are doing things with other tools of the
collection (e.g: after changing certificates in elasticsearch, we need
to restart kibana).
This tasks will be moved to the global role "elasticstack" and can be
called by via 'notify'.
Also, global handlers will only do something when
`elasticstack_collection_managed` is true (means the whole stack is
managed by the collecion)